feat: Node타입 파일 분리
