+++ Calculator/Calculator/Model/CalculatorItemQueue.swift
@@ -42,12 +42,3 @@ class CalculatorItemQueue<T>: CalculateItem {
     }
 }
 
-class Node<T> {
-    var data: T?
-    var next: Node?
-    
-    init(data: T?, next: Node? = nil) {
-        self.data = data
-        self.next = next
-    }
-}
