+++ Calculator/Calculator/Model/Node.swift
@@ -1,0 +1,17 @@
+//
+//  Calculator - ViewController.swift
+//  Created by inho.
+//  Copyright © yagom. All rights reserved.
+//
+
+import Foundation
+
+class Node<T> {
+    var data: T?
+    var next: Node?
+    
+    init(data: T?, next: Node? = nil) {
+        self.data = data
+        self.next = next
+    }
+}
